Some CUDA testing scripts (FFTs and convolutions).
Modified the Zemax output converting script to rotate the FFT PSFs by 180 degree. Zemax 13 does not allow this rotation to be used when outputting to a file while in visualisation it is allowed. It seems that the Huygens PSFs and FFT PSFs are oriented differently when using the Euclid VIS model.