Replaced the ghost model input file. The old one had incorrect sampling.
Added a new function to the object catalogue generation script to generate catalogues with stars drawn from the Bachall and Soneira model.

Simulator now at 1.31. this version has a single FOLDER variable that should be set to point to the vissim-python location so that required parameter files can be found. Also modified the ghost model: the offset from the object is now 750 pixels.